Server/requirements.txt
flask
mysql-connector-python
sqlalchemy
flask_sqlalchemy